+++ rm-datasource/src/test/java/io/seata/rm/datasource/exec/OracleInsertExecutorTest.java
@@ -149,6 +149,19 @@ public class OracleInsertExecutorTest {
 
     }
 
+    @Test
+    public void testGetPkValuesByAuto_NotSupportYetException() {
+        Assertions.assertThrows(NotSupportYetException.class, () -> {
+            doReturn(tableMeta).when(insertExecutor).getTableMeta();
+            Map<String, ColumnMeta> columnMetaMap = new HashMap<>();
+            columnMetaMap.put(ID_COLUMN, new ColumnMeta());
+            columnMetaMap.put(USER_ID_COLUMN, new ColumnMeta());
+            when(tableMeta.getPrimaryKeyMap()).thenReturn(columnMetaMap);
+            insertExecutor.getPkValuesByAuto();
+        });
+    }
+
+
     private List<String> mockInsertColumns() {
         List<String> columns = new ArrayList<>();
         columns.add(ID_COLUMN);
